added tests for login and logout
